+++ rake.d/online_file.rb
@@ -9,7 +9,7 @@ module Mtx::OnlineFile
       file_name ||= url.gsub(%r{.*/}, '')
       file_name   = "tmp/#{file_name}"
 
-      if !FileTest.exists?(file_name)
+      if !FileTest.exist?(file_name)
         @@to_unlink << file_name
 
         runq "wget", url, "wget --quiet -O #{file_name} #{url}"
@@ -30,7 +30,7 @@ module Mtx::OnlineFile
     return if c?(:KEEP_DOWNLOADED_FILES)
 
     @@to_unlink.
-      select { |fn| FileTest.exists? fn }.
+      select { |fn| FileTest.exist? fn }.
       each   { |fn| File.unlink      fn }
   end
 end